backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
google-genai==1.14.0
notion-client==2.3.0
gunicorn==23.0.0
